Refactor: made necessary changes
