+++ pages/contribution-flow.js
@@ -16,7 +16,6 @@ import { STEPS } from '../components/contribution-flow/constants';
 import ContributionBlocker, {
   CONTRIBUTION_BLOCKER,
   getContributionBlocker,
-  Redirect,
 } from '../components/contribution-flow/ContributionBlocker';
 import ContributionFlowSuccess from '../components/contribution-flow/ContributionFlowSuccess';
 import {
@@ -28,6 +27,7 @@ import { getContributionFlowMetadata } from '../components/contribution-flow/uti
 import ErrorPage from '../components/ErrorPage';
 import Loading from '../components/Loading';
 import Page from '../components/Page';
+import Redirect from '../components/Redirect';
 import { withStripeLoader } from '../components/StripeProvider';
 import { withUser } from '../components/UserProvider';
 
@@ -138,10 +138,10 @@ class NewContributionFlowPage extends React.Component {
     }
 
     const contributionBLocker = getContributionBlocker(LoggedInUser, account, tier, Boolean(this.props.tierId));
-    if (contributionBLocker.reason === CONTRIBUTION_BLOCKER.NO_CUSTOM_CONTRIBUTION) {
-      return <Redirect to={`${account.slug}/contribute`} />;
-    }
     if (contributionBLocker) {
+      if (contributionBLocker.reason === CONTRIBUTION_BLOCKER.NO_CUSTOM_CONTRIBUTION) {
+        return <Redirect to={`${account.slug}/contribute`} />;
+      }
       return <ContributionBlocker blocker={contributionBLocker} account={account} />;
     } else if (step === 'success') {
       return <ContributionFlowSuccess collective={account} />;
